updated data type

+++ src/main/java/com/cosmetic/app/repository/UserRepository.java
@@ -6,9 +6,9 @@ import com.cosmetic.app.model.User;
 
 public interface UserRepository {
 	public User saveUser(User user);
-	public User findUserById(Integer id);
+	public User findUserById(long id);
 	public List<User> findAllUsers();
-	public User deleteUser(Integer idUser);
+	public User deleteUser(long idUser);
 	public User updateUser(User user);
 	public boolean validUserAndPassword(String user, String password);
 	public User findUserByName(String username);
